fix: resolve WGSL type errors across 11 variation files plus audit docs
- Fix ternary operators in GPU code → select() calls (wavs4Var, splitVar, xheartVar, floraVar, dustpointVar)
- Fix i32/u32 bit shift type mismatch in circleLinearVar and circleRandVar (u32 cast on shift operand)
- Fix implicit i32→f32 conversion warnings: change i32 struct fields to f32 in invCircle, invCircle2, invEllipse
- Fix i32 comparisons with bare literals in octagonVar (f32), chaosCubesVar (i32 casts), scry2 (i32 cast)
- Add comprehensive JWildfire variation audit CSV (876 variations: 243 exists, 543 not viable, 90 viable)

+++ packages/app/src/flame/variations/parametric/general/chaosCubesVar.tsx
@@ -119,10 +119,10 @@ export const chaosCubesVar = parametricVariation(
     const max_iter = i32(P.depth)
     for (let i = 0; i < max_iter; i++) {
       let cmode = i32(P.mode)
-      if (cmode === 7) {
+      if (cmode === i32(7)) {
         cmode = select(i32(P.mode7_B), i32(P.mode7_A), random() < 0.5)
       }
-      if (cmode < 6) {
+      if (cmode < i32(6)) {
         const rx = floor(random() * 3.0) - 1.0
         const ry = floor(random() * 3.0) - 1.0
         const rz = floor(random() * 3.0) - 1.0
