+++ packages/app/src/flame/variations/parametric/general/circleLinearVar.tsx
@@ -86,7 +86,7 @@ export const circleLinearVar = parametricVariation(
 
     // Inline discretNoise2
     const n1 = i32(M + 10.0) + i32(N + 3.0) * 57
-    const n1x = (n1 << u32(13)) ^ n1
+    const n1x = i32(u32(n1) << u32(13)) ^ n1
     const V =
       (0.3 +
         0.7 *
@@ -95,7 +95,7 @@ export const circleLinearVar = parametricVariation(
       P.sc
 
     const n2 = i32(M + P.seed) + i32(N) * 57
-    const n2x = (n2 << u32(13)) ^ n2
+    const n2x = i32(u32(n2) << u32(13)) ^ n2
     const Z1 =
       f32((n2x * (n2x * n2x * 15731 + 789221) + 1376312589) & 0x7fffffff) * AM
